requirements.txt
shortuuid~=1.0.11
Flask~=2.3.2
SQLAlchemy==2.0.17
Werkzeug~=2.3.6
python-dotenv~=1.0.0